api/requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Django==5.0.1
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
sqlparse==0.4.4
Werkzeug==3.0.1
djangorestframework==3.14.0
pillow==10.2.0
